+++ apps/client/src/app/app.component.ts
@@ -3,11 +3,7 @@ import { HoldingDetailDialogParams } from '@ghostfolio/client/components/holding
 import { getCssVariable } from '@ghostfolio/common/helper';
 import { InfoItem, User } from '@ghostfolio/common/interfaces';
 import { hasPermission, permissions } from '@ghostfolio/common/permissions';
-import {
-  internalRoutes,
-  publicRoutes,
-  routes
-} from '@ghostfolio/common/routes/routes';
+import { internalRoutes, publicRoutes } from '@ghostfolio/common/routes/routes';
 import { ColorScheme } from '@ghostfolio/common/types';
 
 import { DOCUMENT } from '@angular/common';
@@ -202,7 +198,7 @@ export class AppComponent implements OnDestroy, OnInit {
             this.currentRoute === publicRoutes.faq.path ||
             this.currentRoute === publicRoutes.resources.path ||
             this.currentRoute === internalRoutes.account.path ||
-            this.currentRoute === routes.adminControl ||
+            this.currentRoute === internalRoutes.adminControl.path ||
             this.currentRoute === internalRoutes.home.path ||
             this.currentRoute === internalRoutes.portfolio.path ||
             this.currentRoute === internalRoutes.zen.path) &&
@@ -213,10 +209,10 @@ export class AppComponent implements OnDestroy, OnInit {
             this.currentRoute === publicRoutes.features.path ||
             this.currentRoute === publicRoutes.markets.path ||
             this.currentRoute === publicRoutes.openStartup.path ||
-            this.currentRoute === routes.public ||
+            this.currentRoute === publicRoutes.public.path ||
             this.currentRoute === publicRoutes.pricing.path ||
             this.currentRoute === publicRoutes.register.path ||
-            this.currentRoute === routes.start) &&
+            this.currentRoute === publicRoutes.start.path) &&
           this.deviceType !== 'mobile';
 
         if (this.deviceType === 'mobile') {