+++ apps/client/src/app/pages/admin/admin-page.component.ts
@@ -1,5 +1,5 @@
 import { TabConfiguration } from '@ghostfolio/common/interfaces';
-import { routes } from '@ghostfolio/common/routes/routes';
+import { internalRoutes } from '@ghostfolio/common/routes/routes';
 
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { DeviceDetectorService } from 'ngx-device-detector';
@@ -27,27 +27,27 @@ export class AdminPageComponent implements OnDestroy, OnInit {
       {
         iconName: 'reader-outline',
         label: $localize`Overview`,
-        routerLink: ['/' + routes.adminControl]
+        routerLink: internalRoutes.adminControl.routerLink
       },
       {
         iconName: 'settings-outline',
-        label: $localize`Settings`,
-        routerLink: ['/' + routes.adminControl, routes.settings]
+        label: internalRoutes.adminControl.subRoutes.settings.title,
+        routerLink: internalRoutes.adminControl.subRoutes.settings.routerLink
       },
       {
         iconName: 'server-outline',
-        label: $localize`Market Data`,
-        routerLink: ['/' + routes.adminControl, routes.marketData]
+        label: internalRoutes.adminControl.subRoutes.marketData.title,
+        routerLink: internalRoutes.adminControl.subRoutes.marketData.routerLink
       },
       {
         iconName: 'flash-outline',
-        label: $localize`Job Queue`,
-        routerLink: ['/' + routes.adminControl, routes.jobs]
+        label: internalRoutes.adminControl.subRoutes.jobs.title,
+        routerLink: internalRoutes.adminControl.subRoutes.jobs.routerLink
       },
       {
         iconName: 'people-outline',
-        label: $localize`Users`,
-        routerLink: ['/' + routes.adminControl, routes.users]
+        label: internalRoutes.adminControl.subRoutes.users.title,
+        routerLink: internalRoutes.adminControl.subRoutes.users.routerLink
       }
     ];
   }