+++ apps/client/src/app/pages/user-account/user-account-page.component.ts
@@ -1,6 +1,6 @@
 import { UserService } from '@ghostfolio/client/services/user/user.service';
 import { TabConfiguration, User } from '@ghostfolio/common/interfaces';
-import { internalRoutes, routes } from '@ghostfolio/common/routes/routes';
+import { internalRoutes } from '@ghostfolio/common/routes/routes';
 
 import { ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
 import { DeviceDetectorService } from 'ngx-device-detector';
@@ -39,17 +39,15 @@ export class UserAccountPageComponent implements OnDestroy, OnInit {
             },
             {
               iconName: 'diamond-outline',
-              label: $localize`Membership`,
-              routerLink: [
-                '/' + internalRoutes.account.path,
-                routes.membership
-              ],
+              label: internalRoutes.account.subRoutes.membership.title,
+              routerLink:
+                internalRoutes.account.subRoutes.membership.routerLink,
               showCondition: !!this.user?.subscription
             },
             {
               iconName: 'key-outline',
-              label: $localize`Access`,
-              routerLink: ['/' + internalRoutes.account.path, routes.access]
+              label: internalRoutes.account.subRoutes.access.title,
+              routerLink: internalRoutes.account.subRoutes.access.routerLink
             }
           ];